Permission Fixes

+++ commands/administration/createMemberChannel.js
@@ -7,7 +7,9 @@ module.exports = {
 
     async execute(message,args,client){
         if(!message.member.permissions.has('ADMINSTRATOR')) return message.channel.send({content:'Perms Denied'})
-
+        if(!message.guild.me.permissions.has('MANAGE_GUILD') || message.guild.me.permissions.has('MANAGE_MESSAGES')) {
+            return message.channel.send({content:'I must have ADMIN command to run this command'})
+        }
         Schema.findOne({Guild : message.guild.id}, async(err, data) => {
             try{
             if(data) data.delete()
